+++ packages/search-in-workspace/src/browser/search-in-workspace-frontend-module.ts
@@ -27,6 +27,7 @@ import { SearchInWorkspaceFrontendContribution } from './search-in-workspace-fro
 import { InMemoryTextResourceResolver } from './in-memory-text-resource';
 import { SearchInWorkspaceContextKeyService } from './search-in-workspace-context-key-service';
 import { TabBarToolbarContribution } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
+import { bindSearchInWorkspacePreferences } from './search-in-workspace-preferences';
 
 export default new ContainerModule(bind => {
     bind(SearchInWorkspaceContextKeyService).toSelf().inSingletonScope();
@@ -55,6 +56,8 @@ export default new ContainerModule(bind => {
 
     bind(InMemoryTextResourceResolver).toSelf().inSingletonScope();
     bind(ResourceResolver).toService(InMemoryTextResourceResolver);
+
+    bindSearchInWorkspacePreferences(bind);
 });
 
 export function createSearchTreeWidget(parent: interfaces.Container): SearchInWorkspaceResultTreeWidget {